+++ kern/hub.cpp
@@ -60,7 +60,7 @@ std::vector<std::string> Hub::GenerateNetDeviceCpp()
 {
   std::vector<std::string> ndc;
   std::vector<std::string> allNodes = this->GroupAsNodeContainerCpp();
-  for(size_t i = 0; i < (size_t) allNodes.size(); i++)
+  for(size_t i = 0; i <  allNodes.size(); i++)
   {
     ndc.push_back(allNodes.at(i));
   }
@@ -103,7 +103,7 @@ std::vector<std::string> Hub::GenerateNetDevicePython()
   std::vector<std::string> ret;
   std::vector<std::string> allNodes = this->GroupAsNodeContainerPython();
   
-  for(size_t i = 0; i < (size_t) allNodes.size(); i++)
+  for(size_t i = 0; i <  allNodes.size(); i++)
   {
     ret.push_back(allNodes.at(i));
   }
