+++ kern/point-to-point.cpp
@@ -60,7 +60,7 @@ std::vector<std::string> PointToPoint::GenerateNetDeviceCpp()
 {
   std::vector<std::string> ndc;
   std::vector<std::string> allNodes = this->GroupAsNodeContainerCpp();
-  for(size_t i = 0; i < (size_t) allNodes.size(); i++)
+  for(size_t i = 0; i <  allNodes.size(); i++)
   {
     ndc.push_back(allNodes.at(i));
   }
@@ -96,7 +96,7 @@ std::vector<std::string> PointToPoint::GenerateNetDevicePython()
   std::vector<std::string> ndc;
   std::vector<std::string> allNodes = this->GroupAsNodeContainerPython();
   
-  for(size_t i = 0; i < (size_t) allNodes.size(); i++)
+  for(size_t i = 0; i <  allNodes.size(); i++)
   {
     ndc.push_back(allNodes.at(i));
   }
